Changes Committed

+++ Code4.java
@@ -1,8 +1,9 @@
+//Typecast 2
 public class Code4 {
     public static void main(String...args){
         int a=29;
         float b=48f;
-        float sum=a+b; // We cannot Convert from float to int in java 
+        int sum=(int) ((int)a+b); // We cannot Convert from float to int in java 
         System.out.println(sum);
         System.out.println(b-a);
         System.out.println(b/a);
